split out app content. Also receive feedback as an additional comment instead of a replacement message
